+++ sbi-spec/CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Added
 
+- base: add Coreboot and Oreboot to `impl_id` module
+
 ### Modified
 
 ### Fixed